5.110 Initial Commit (#64)
Talents, Powers, Disciplines, Crafting Recipes, and Promotions have all been included in this build. Lots of structural data changes that are going to blow up the changelog system, but oh well. 5.110 has tons of structural changes.

Source information is now kept on the source, instead of being a field on the destination. For example;

Disciplines -> point to -> Powers
instead of
Powers -> pointing to -> Disciplines.

This will have some ramifications on API data but is necessary due to the way we are now gathering data directly from client.

Note: Changefeed will false-positive a ton with some id's being changed based on how the data appears in the crowfall client. Expect lots of "deleted" and "added" with this version.